slight refactors
